feat: set interface naming conventions on eslint
