+++ .eslintrc.json
@@ -23,10 +23,6 @@
       "error",
       2
     ],
-    "linebreak-style": [
-      "error",
-      "windows"
-    ],
     "quotes": [
       "error",
       "single"
@@ -34,6 +30,19 @@
     "semi": [
       "error",
       "always"
+    ],
+    "@typescript-eslint/naming-convention": [
+      "error",
+      {
+        "selector": "interface",
+        "format": [
+          "PascalCase"
+        ],
+        "custom": {
+          "regex": "^I[A-Z]",
+          "match": true
+        }
+      }
     ]
   }
 }
