--- src/types/IFilter.ts
@@ -1,5 +1,0 @@
-export interface IFilter {
-  name: string;
-  isActive: boolean;
-  id: number;
-}
