--- src/types/Modal.ts
@@ -1,6 +1,0 @@
-export interface IModal {
-  open: boolean;
-  onClose: () => void;
-  title: string;
-  children?: React.ReactNode;
-}
