+++ src/types/interfaces/IFilter.ts
@@ -1,0 +1,5 @@
+export interface IFilter {
+  name: string;
+  isActive: boolean;
+  id: number;
+}
