+++ src/types/interfaces/IModal.ts
@@ -1,0 +1,6 @@
+export interface IModal {
+  open: boolean;
+  onClose: () => void;
+  title: string;
+  children?: React.ReactNode;
+}
